pyproject.toml
[tool.poetry]
name = "lyric-generation"
version = "0.1.0"
description = ""
authors = ["tanikawa04 <tanikawa04@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "1.18.*"
torch = "1.4.*"
sudachipy = "0.4.*"
sudachidict-core = {url = "https://object-storage.tyo2.conoha.io/v1/nc_2520839e1f9641b08211a5c85243124a/sudachi/SudachiDict_core-20200330.tar.gz"}
sentencepiece = "^0.1.86"
click = "^7.1.2"
pytorch-ignite = "^0.3.0"
tensorboard = "^2.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
